added ground check to player movement

+++ Assets/harley/scripes/Inerface.cs
@@ -3,5 +3,5 @@ using System.Collections.Generic;
 using UnityEngine;
 public interface IDamageable
 {
-    void takeDamage(int damageTaken,int pointgain, player_movment player);
+    void takeDamage(float damageTaken,int pointgain, player_movment player);
 }
